test(core): isolate EA process scenarios
Run process-tree timing tests in the nonparallel runner collection and publish helper launch records atomically with bounded reads.

Refs #2104

Copilot CLI: 1.0.80
Model: GPT-5.6 Sol (gpt-5.6-sol)
